+++ lib/azure/storage/file/file_service.rb
@@ -35,7 +35,7 @@ module Azure::Storage
 
       def initialize(options = {}, &block)
         client_config = options[:client] || Azure::Storage
-        signer = options[:signer] || Core::Auth::SharedKey.new(client_config.storage_account_name, client_config.storage_access_key)
+        signer = options[:signer] || client_config.signer || Core::Auth::SharedKey.new(client_config.storage_account_name, client_config.storage_access_key)
         super(signer, client_config.storage_account_name, options, &block)
         @host = client.storage_file_host
       end